+++ iOS/Abra/Controllers/OfferSelect/OSTimeViewController.swift
@@ -8,43 +8,6 @@
 
 import UIKit
 
-enum OSTimeMode {
-    case origin
-    case destination
-}
-
-protocol OSTimeDelegate {
-    func `continue`(time: Date)
-}
-
-class OSTimeViewModel: NSObject {
-    
-    var delegate: OSTimeDelegate?
-    var timeMode: OSTimeMode = .origin {
-        didSet {
-            switch timeMode {
-            case .origin:
-                switch OSBaseViewController.offerSelectMode {
-                case .courier:
-                    titleLabel.text = "What time would you like to send?"
-                case .carry:
-                    titleLabel.text = "What time are you going?"
-                }
-            case .destination:
-                titleLabel.text = "What time are you arriving?"
-            }
-        }
-    }
-    
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var datePicker: UIDatePicker!
-    
-    @IBAction func `continue`(_ sender: UIButton) {
-        delegate?.`continue`(time: datePicker.date)
-    }
-    
-}
-
 class OSTimeViewController: OSBaseViewController {
     
     var timeMode: OSTimeMode = .origin
