+++ iOS/Abra/Controllers/Shared/BaseNavigationController.swift
@@ -41,7 +41,7 @@ class BaseNavigationController: UINavigationController {
     }
     
     override var preferredStatusBarStyle: UIStatusBarStyle {
-        return .default
+        return .`default`
     }
     
 }
